added scroll to top to search page
