+++ src/components/ScrollToTop/ScrollToTop.styled.js
@@ -1,0 +1,9 @@
+import styled from 'styled-components';
+
+export const ScrollToTopImg = styled.img`
+  position: fixed;
+  right: 10px;
+  bottom: 34px;
+  width: 3rem;
+  height: initial;
+`;
